second commit

+++ ft_memset.c
@@ -1,4 +1,18 @@
-void	*ft_memset(void *b, int c, unsigned int len)
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        ::::::::            */
+/*   ft_memset.c                                        :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: pniezen <pniezen@student.codam.nl>           +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2021/10/19 16:34:37 by pniezen       #+#    #+#                 */
+/*   Updated: 2021/10/20 15:26:51 by pniezen       ########   odam.nl         */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "libft.h"
+
+void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char	*p;
